src/requirements.txt
Flask
pyjwt
mysql-connector-python
python-dotenv
